New ACL mechanism including examples.

+++ discussment-core/src/main/java/org/danekja/discussment/core/service/imp/DefaultCategoryService.java
@@ -1,42 +1,71 @@
 package org.danekja.discussment.core.service.imp;
 
+import org.danekja.discussment.core.accesscontrol.domain.AccessDeniedException;
+import org.danekja.discussment.core.accesscontrol.domain.Action;
+import org.danekja.discussment.core.accesscontrol.domain.IDiscussionUser;
+import org.danekja.discussment.core.accesscontrol.domain.PermissionType;
+import org.danekja.discussment.core.accesscontrol.service.AccessControlService;
+import org.danekja.discussment.core.accesscontrol.service.DiscussionUserService;
 import org.danekja.discussment.core.dao.CategoryDao;
 import org.danekja.discussment.core.domain.Category;
 import org.danekja.discussment.core.service.CategoryService;
 
 import java.util.List;
 
-/**
- * Created by Martin Bláha on 28.01.17.
- */
 public class DefaultCategoryService implements CategoryService {
 
     private CategoryDao categoryDao;
+    private AccessControlService accessControlService;
+    private DiscussionUserService discussionUserService;
 
-    public DefaultCategoryService(CategoryDao categoryDao) {
+    public DefaultCategoryService(CategoryDao categoryDao, AccessControlService accessControlService, DiscussionUserService discussionUserService) {
         this.categoryDao = categoryDao;
+        this.accessControlService = accessControlService;
+        this.discussionUserService = discussionUserService;
     }
 
-    public Category createCategory(Category entity) {
-
-        return categoryDao.save(entity);
-
+    @Override
+    public Category createCategory(Category entity) throws AccessDeniedException {
+        if(accessControlService.canAddCategory()) {
+            return categoryDao.save(entity);
+        } else {
+            throw new AccessDeniedException(Action.CREATE, getCurrentUserId(), 0, PermissionType.CATEGORY);
+        }
     }
 
-    public Category getCategoryById(long categoryId) {
-
-        return categoryDao.getById(categoryId);
+    @Override
+    public Category getCategoryById(long categoryId) throws AccessDeniedException {
+        if (accessControlService.canViewCategories()) {
+            return categoryDao.getById(categoryId);
+        } else {
+            throw new AccessDeniedException(Action.VIEW, getCurrentUserId(), categoryId, PermissionType.CATEGORY);
+        }
     }
 
-    public List<Category> getCategories() {
-
-        return categoryDao.getCategories();
+    @Override
+    public List<Category> getCategories() throws AccessDeniedException {
+        if (accessControlService.canViewCategories()) {
+            return categoryDao.getCategories();
+        } else {
+            throw new AccessDeniedException(Action.VIEW, getCurrentUserId(), 0, PermissionType.CATEGORY);
+        }
     }
 
-    public void removeCategory(Category entity) {
-
-        categoryDao.remove(entity);
+    @Override
+    public void removeCategory(Category entity) throws AccessDeniedException {
+        if (accessControlService.canRemoveCategory(entity)) {
+            categoryDao.remove(entity);
+        } else {
+            throw new AccessDeniedException(Action.DELETE, getCurrentUserId(), entity.getId(), PermissionType.CATEGORY);
+        }
     }
 
-
+    /**
+     * Returns the id of the currently logged user. Used when throwing access denied exception.
+     * @return Id of the currently logged user or null if no user is logged.
+     */
+    private String getCurrentUserId() {
+        IDiscussionUser user = discussionUserService.getCurrentlyLoggedUser();
+        return user == null ? null : user.getDiscussionUserId();
+    }
 }
